+++ src/App.js
@@ -4,7 +4,7 @@ import { MuiThemeProvider } from '@material-ui/core/styles'
 import { injectGlobal } from 'emotion'
 import { Navigation } from 'components/shared/navigation'
 import Breadcrumbs from 'components/shared/navigation/breadcrumbs/Breadcrumbs'
-import Notifications from 'containers/notification-service/NotificationService'
+import Notifications from 'containers/NotificationService'
 import React from 'react'
 import styled from 'react-emotion'
 import { ThemeProvider } from 'emotion-theming'
