+++ src/routes/ClaimsPageRoute.tsx
@@ -1,5 +1,5 @@
-import ClaimDetailsPage from 'containers/claim-details-page/ClaimDetailsPage'
-import { ClaimsPage } from 'containers/claims-page/ClaimsPage'
+import ClaimDetailsPage from 'containers/ClaimDetailsPage'
+import { ClaimsPage } from 'containers/ClaimsPage'
 import * as PropTypes from 'prop-types'
 import * as React from 'react'
 import { Route, Switch } from 'react-router'
