+++ src/routes/LoginPageRoute.tsx
@@ -1,4 +1,4 @@
-import LoginPage from 'containers/login-page/LoginPage'
+import LoginPage from 'containers/LoginPage'
 import * as React from 'react'
 
 const LoginPageRouter = () => <LoginPage />
