+++ src/store/actions/clientActions.js
@@ -1,8 +1,8 @@
-import {CLIENT_SET, CLIENT_UNSET} from '../constants/login'
+import { CLIENT_SET, CLIENT_UNSET } from '../constants/login'
 
-export const setClient = (creditals) => ({
+export const setClient = (credentials) => ({
   type: CLIENT_SET,
-  creditals,
+  credentials,
 })
 
 export const unsetClient = () => ({
